PyTest DQ Framework/requirements.txt
pytest
pytest-html
psycopg2-binary
pandas
